writable6

+++ src/main/java/ca/uwaterloo/cs451/a1/PairsPMI.java
@@ -146,7 +146,7 @@ public class PairsPMI  extends Configured implements Tool {
 
   private static final class MyCombiner extends
       Reducer<PairOfStrings, FloatWritable, PairOfStrings, FloatWritable> {
-    private static final FloatWritable SUM = new IntWritable();
+    private static final FloatWritable SUM = new FloatWritable();
 
     @Override
     public void reduce(PairOfStrings key, Iterable<FloatWritable> values, Context context)
